feat: rename SimpleCacheClient to CacheClient
This commit renames the primary client object from SimpleCacheClient
to CacheClient, after we received feedback from users that the
name was misleading and seemed to imply that there were multiple
clients to choose from.

The legacy name is retained as a type alias with a deprecation
warning so that this change is backward-compatible.
