+++ src/Momento.Sdk/ICacheClient.cs
@@ -12,7 +12,7 @@ namespace Momento.Sdk;
 ///
 /// Includes control operations and data operations.
 /// </summary>
-public interface ISimpleCacheClient : IDisposable
+public interface ICacheClient : IDisposable
 {
     /// <summary>
     /// Creates a cache if it doesn't exist.
