+++ tests/Integration/Momento.Sdk.Tests/SetTest.cs
@@ -6,10 +6,10 @@ using Momento.Sdk.Tests;
 
 namespace Momento.Sdk.Tests;
 
-[Collection("SimpleCacheClient")]
+[Collection("CacheClient")]
 public class SetTest : TestBase
 {
-    public SetTest(SimpleCacheClientFixture fixture) : base(fixture)
+    public SetTest(CacheClientFixture fixture) : base(fixture)
     {
     }
